fix: update dependencies and go version
The go version has been updated to the same version that is used in the latest caddy:2.7-builder
docker image.

The AWS SDK and Caddy go dependencies also have been updated to the latest versions. This fixes an
unknown error with the generated signature that resulted in a forbidden response from AWS.

With these changes the module works as intended, at least for my use case (IAM-authorizer protected
route in AWS API Gateway). I would assume that the signing works for other AWS Sigv4 protected
endpoints as well.